fix error to allow filter_hot_events=False (b5bb4a8)
The code for hot pixel filtering introduced in b5bb4a8 seems to break if filter_hot_events=False. In this case, self.hot_events_mask is a numpy array that is later multiplied with a torch tensor (data_loader.dataset, line 328). This seems to be an appropriate fix.

Hope this is helpful.